# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ProofForge API
  description: Propositions, arguments, cohérence, stabilité et couverture d'obligations
    de preuves immuables avec vérification machine déterministe.
  version: 1.0.7
paths:
...
  /v1/propositions:
    get:
      summary: List Propositions
      operationId: list_propositions_v1_propositions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Proposition'
                type: array
                title: Response List Propositions V1 Propositions Get
    post:
      summary: Create Proposition
      operationId: create_proposition_v1_propositions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropositionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposition'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/verifications:
    post:
      summary: Create Verification
      operationId: create_verification_v1_verifications_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Verification'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/arguments:
    get:
      summary: List Arguments
      operationId: list_arguments_v1_arguments_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Argument'
                type: array
                title: Response List Arguments V1 Arguments Get
    post:
      summary: Create Argument
      operationId: create_argument_v1_arguments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArgumentCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Argument'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/entailments:
    post:
      summary: Check Entailment
      operationId: check_entailment_v1_entailments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntailmentCheckCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntailmentCheck'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/formula-comparisons:
    get:
      summary: List Formula Comparisons
      operationId: list_formula_comparisons_v1_formula_comparisons_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FormulaComparison'
                type: array
                title: Response List Formula Comparisons V1 Formula Comparisons Get
    post:
      summary: Compare Formulae
      operationId: compare_formulae_v1_formula_comparisons_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormulaComparisonCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormulaComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/formula-comparisons/{comparison_id}:
    get:
      summary: Get Formula Comparison
      operationId: get_formula_comparison_v1_formula_comparisons__comparison_id__get
      parameters:
      - name: comparison_id
        in: path
        required: true
        schema:
          type: string
          title: Comparison Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormulaComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/premise-sets:
    get:
      summary: List Premise Sets
      operationId: list_premise_sets_v1_premise_sets_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PremiseSet'
                type: array
                title: Response List Premise Sets V1 Premise Sets Get
    post:
      summary: Create Premise Set
      operationId: create_premise_set_v1_premise_sets_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PremiseSetCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PremiseSet'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/inconsistency-analyses:
    post:
      summary: Analyze Inconsistency
      operationId: analyze_inconsistency_v1_inconsistency_analyses_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InconsistencyAnalysisCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InconsistencyAnalysis'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/inconsistency-analyses/{analysis_id}:
    get:
      summary: Get Inconsistency Analysis
      operationId: get_inconsistency_analysis_v1_inconsistency_analyses__analysis_id__get
      parameters:
      - name: analysis_id
        in: path
        required: true
        schema:
          type: string
          title: Analysis Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InconsistencyAnalysis'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/proof-dependency-dossiers:
    get:
      summary: List Proof Dependency Dossiers
      operationId: list_proof_dependency_dossiers_v1_proof_dependency_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProofDependencyDossier'
                type: array
                title: Response List Proof Dependency Dossiers V1 Proof Dependency
                  Dossiers Get
    post:
      summary: Create Proof Dependency Dossier
      operationId: create_proof_dependency_dossier_v1_proof_dependency_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofDependencyDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofDependencyDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/proof-dependency-dossiers/{dossier_id}:
    get:
      summary: Get Proof Dependency Dossier
      operationId: get_proof_dependency_dossier_v1_proof_dependency_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofDependencyDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/multi-argument-coherence-dossiers:
    get:
      summary: List Multi Argument Coherence Dossiers
      operationId: list_multi_argument_coherence_dossiers_v1_multi_argument_coherence_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MultiArgumentCoherenceDossier'
                type: array
                title: Response List Multi Argument Coherence Dossiers V1 Multi Argument
                  Coherence Dossiers Get
    post:
      summary: Create Multi Argument Coherence Dossier
      operationId: create_multi_argument_coherence_dossier_v1_multi_argument_coherence_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiArgumentCoherenceCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiArgumentCoherenceDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/multi-argument-coherence-dossiers/{dossier_id}:
    get:
      summary: Get Multi Argument Coherence Dossier
      operationId: get_multi_argument_coherence_dossier_v1_multi_argument_coherence_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiArgumentCoherenceDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/proof-stability-dossiers:
    get:
      summary: List Proof Stability Dossiers
      operationId: list_proof_stability_dossiers_v1_proof_stability_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProofStabilityDossier'
                type: array
                title: Response List Proof Stability Dossiers V1 Proof Stability Dossiers
                  Get
    post:
      summary: Create Proof Stability Dossier
      operationId: create_proof_stability_dossier_v1_proof_stability_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofStabilityDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofStabilityDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/proof-stability-dossiers/{dossier_id}:
    get:
      summary: Get Proof Stability Dossier
      operationId: get_proof_stability_dossier_v1_proof_stability_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofStabilityDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/proof-obligation-coverage-dossiers:
    get:
      summary: List Proof Obligation Coverage Dossiers
      operationId: list_proof_obligation_coverage_dossiers_v1_proof_obligation_coverage_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProofObligationCoverageDossier'
                type: array
                title: Response List Proof Obligation Coverage Dossiers V1 Proof Obligation
                  Coverage Dossiers Get
    post:
      summary: Create Proof Obligation Coverage Dossier
      operationId: create_proof_obligation_coverage_dossier_v1_proof_obligation_coverage_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofObligationCoverageDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofObligationCoverageDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/proof-obligation-coverage-dossiers/{dossier_id}:
    get:
      summary: Get Proof Obligation Coverage Dossier
      operationId: get_proof_obligation_coverage_dossier_v1_proof_obligation_coverage_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofObligationCoverageDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Argument:
      properties:
        id:
          type: string
          title: Id
        argument_hash:
          type: string
          title: Argument Hash
        specification:
          $ref: '#/components/schemas/ArgumentCreate'
        immutable:
          type: boolean
          title: Immutable
          default: true
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - argument_hash
      - specification
      - created_at
      title: Argument
    ArgumentCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        premises:
          items:
            type: string
          type: array
          maxItems: 32
          minItems: 1
          title: Premises
        conclusion:
          type: string
          maxLength: 1000
          minLength: 1
          title: Conclusion
        variables:
          items:
            type: string
          type: array
          maxItems: 8
          minItems: 1
          title: Variables
        language:
          type: string
          pattern: ^propositional-v1$
          title: Language
          default: propositional-v1
      additionalProperties: false
      type: object
      required:
      - name
      - premises
      - conclusion
      - variables
      title: ArgumentCreate
...
    EntailmentCheck:
      properties:
        id:
          type: string
          title: Id
        argument_id:
          type: string
          title: Argument Id
        method:
          type: string
          title: Method
        verdict:
          type: string
          enum:
          - ENTAILED
          - NOT_ENTAILED
          - INCONSISTENT_PREMISES
          title: Verdict
        valuations_checked:
          type: integer
          title: Valuations Checked
        premise_models:
          type: integer
          title: Premise Models
        counterexample:
          anyOf:
          - additionalProperties:
              type: boolean
            type: object
          - type: 'null'
          title: Counterexample
        entailment_hash:
          type: string
          title: Entailment Hash
        reproducible:
          type: boolean
          title: Reproducible
          default: true
        idempotent_replay:
          type: boolean
          title: Idempotent Replay
          default: false
        immutable:
          type: boolean
          title: Immutable
          default: true
        created_at:
          type: string
          title: Created At
        warning:
          type: string
          title: Warning
          default: Entaillement limité au langage formel déclaré ; il ne valide ni
            les prémisses dans le monde réel ni une vérité empirique.
      type: object
      required:
      - id
      - argument_id
      - method
      - verdict
      - valuations_checked
      - premise_models
      - entailment_hash
      - created_at
      title: EntailmentCheck
    EntailmentCheckCreate:
      properties:
        argument_id:
          type: string
          minLength: 1
          title: Argument Id
        method:
          type: string
          const: truth-table-entailment-v1
          title: Method
          default: truth-table-entailment-v1
      additionalProperties: false
      type: object
      required:
      - argument_id
      title: EntailmentCheckCreate
    FormulaComparison:
      properties:
        id:
          type: string
          title: Id
        input_hash:
          type: string
          title: Input Hash
        specification:
          $ref: '#/components/schemas/FormulaComparisonCreate'
        verdict:
          type: string
          enum:
          - EQUIVALENT
          - NOT_EQUIVALENT
          title: Verdict
        valuations_checked:
          type: integer
          title: Valuations Checked
        counterexample:
          anyOf:
          - $ref: '#/components/schemas/FormulaCounterexample'
          - type: 'null'
        comparison_hash:
          type: string
          title: Comparison Hash
        reproducible:
          type: boolean
          title: Reproducible
          default: true
        idempotent_replay:
          type: boolean
          title: Idempotent Replay
          default: false
        immutable:
          type: boolean
          title: Immutable
          default: true
        created_at:
          type: string
          title: Created At
        warning:
          type: string
          title: Warning
          default: Équivalence limitée au langage formel déclaré ; elle ne démontre
            aucune équivalence empirique entre les phénomènes éventuellement représentés.
      type: object
      required:
      - id
      - input_hash
      - specification
      - verdict
      - valuations_checked
      - comparison_hash
      - created_at
      title: FormulaComparison
    FormulaComparisonCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        left_expression:
          type: string
          maxLength: 1000
          minLength: 1
          title: Left Expression
        right_expression:
          type: string
          maxLength: 1000
          minLength: 1
          title: Right Expression
        variables:
          items:
            type: string
          type: array
          maxItems: 8
          minItems: 1
          title: Variables
        method:
          type: string
          const: truth-table-equivalence-v1
          title: Method
          default: truth-table-equivalence-v1
      additionalProperties: false
      type: object
      required:
      - name
      - left_expression
      - right_expression
      - variables
      title: FormulaComparisonCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PremiseSet:
      properties:
        id:
          type: string
          title: Id
        premise_set_hash:
          type: string
          title: Premise Set Hash
        specification:
          $ref: '#/components/schemas/PremiseSetCreate'
        immutable:
          type: boolean
          title: Immutable
          default: true
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - premise_set_hash
      - specification
      - created_at
      title: PremiseSet
    PremiseSetCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        premises:
          items:
            type: string
          type: array
          maxItems: 32
          minItems: 1
          title: Premises
        variables:
          items:
            type: string
          type: array
          maxItems: 8
          minItems: 1
          title: Variables
        language:
          type: string
          const: propositional-v1
          title: Language
          default: propositional-v1
      additionalProperties: false
      type: object
      required:
      - name
      - premises
      - variables
      title: PremiseSetCreate
...
    Proposition:
      properties:
        id:
          type: string
          title: Id
        proposition_hash:
          type: string
          title: Proposition Hash
        specification:
          $ref: '#/components/schemas/PropositionCreate'
        immutable:
          type: boolean
          title: Immutable
          default: true
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - proposition_hash
      - specification
      - created_at
      title: Proposition
    PropositionCreate:
      properties:
        name:
          type: string
          minLength: 1
          title: Name
        statement:
          type: string
          minLength: 1
          title: Statement
        formal_expression:
          type: string
          maxLength: 1000
          minLength: 1
          title: Formal Expression
        variables:
          items:
            type: string
          type: array
          maxItems: 8
          minItems: 1
          title: Variables
        language:
          type: string
          pattern: ^propositional-v1$
          title: Language
          default: propositional-v1
      additionalProperties: false
      type: object
      required:
      - name
      - statement
      - formal_expression
      - variables
      title: PropositionCreate
...
    Verification:
      properties:
        id:
          type: string
          title: Id
        proposition_id:
          type: string
          title: Proposition Id
        method:
          type: string
          title: Method
        verdict:
          type: string
          title: Verdict
        valuations_checked:
          type: integer
          title: Valuations Checked
        counterexample:
          anyOf:
          - additionalProperties:
              type: boolean
            type: object
          - type: 'null'
          title: Counterexample
        verification_hash:
          type: string
          title: Verification Hash
        reproducible:
          type: boolean
          title: Reproducible
        idempotent_replay:
          type: boolean
          title: Idempotent Replay
          default: false
        created_at:
          type: string
          title: Created At
        warning:
          type: string
          title: Warning
          default: Verdict limité au langage formel vérifié ; aucune vérité empirique
            n’est déduite.
      type: object
      required:
      - id
      - proposition_id
      - method
      - verdict
      - valuations_checked
      - verification_hash
      - reproducible
      - created_at
      title: Verification
    VerificationCreate:
      properties:
        proposition_id:
          type: string
          title: Proposition Id
        method:
          type: string
          pattern: ^truth-table-v1$
          title: Method
          default: truth-table-v1
      additionalProperties: false
      type: object
      required:
      - proposition_id
      title: VerificationCreate